+++ assets/js/main.js
@@ -5,10 +5,16 @@
  */
 const currentTheme = localStorage.getItem("theme");
 const toggleSwitch = document.querySelector('input[type="checkbox"]');
-const { DarkReader } = window;
+const darkreader = require("darkreader");
+
+const hljs = require("highlight.js/lib/core");
+
+hljs.registerLanguage("python", require("highlight.js/lib/languages/python"));
+
+hljs.highlightAll();
 
 // noinspection JSUnresolvedFunction
-DarkReader.setFetchMethod(window.fetch);
+darkreader.setFetchMethod(window.fetch);
 
 // trigger on import
 // determine the saved state (if there is one)
@@ -16,7 +22,7 @@ if (currentTheme) {
   document.documentElement.setAttribute("data-theme", currentTheme);
   if (currentTheme === "dark") {
     toggleSwitch.checked = true;
-    DarkReader.enable({
+    darkreader.enable({
       brightness: 100,
       contrast: 90,
       sepia: 10,
@@ -57,7 +63,7 @@ function toggleDarkReader() {
   if (toggleSwitch.checked === true) {
     document.documentElement.setAttribute("data-theme", "dark");
     localStorage.setItem("theme", "dark");
-    DarkReader.enable({
+    darkreader.enable({
       brightness: 100,
       contrast: 90,
       sepia: 10,
@@ -65,9 +71,12 @@ function toggleDarkReader() {
   } else {
     document.documentElement.setAttribute("data-theme", "light");
     localStorage.setItem("theme", "light");
-    DarkReader.disable();
+    darkreader.disable();
   }
 }
 
 // default to false if no saved state
 toggleSwitch.addEventListener("change", toggleDarkReader, false);
+
+// noinspection JSUnusedGlobalSymbols
+module.exports = { toggleDarkReader, setMessageCount, setTaskProgress };